refactor(weather): Move weather screens to UI layer
- Relocated city_search_screen.dart, error_display_screen.dart, and weather_result_screen.dart to ui/weather/screens/
- Updated imports to reflect new widget and provider paths
- Promotes better separation between presentation and logic layers
