+++ lib/ui/weather/screens/city_search_screen.dart
@@ -3,9 +3,9 @@ import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 
-import '../../components/background_image.dart';
-import '../../components/city_search_button.dart';
-import '../../components/city_search_input.dart';
+import '../../../../ui/weather/widgets/background_image.dart';
+import '../../../../ui/weather/widgets/city_search_button.dart';
+import '../../../../ui/weather/widgets/city_search_input.dart';
 
 /// [CitySearchScreen]は、ユーザーが都市名を入力して天気情報を検索する画面です。
 class CitySearchScreen extends ConsumerWidget {
