+++ lib/ui/weather/screens/error_display_screen.dart
@@ -3,9 +3,9 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 
 import '../../../../core/extensions/api_error_ui_message.dart';
-import '../../components/app_back_button.dart';
-import '../../components/background_image.dart';
-import '../../view_model/providers/error_view_model_provider.dart';
+import '../../../../ui/weather/view_model/providers/error_view_model_provider.dart';
+import '../../../../ui/weather/widgets/app_back_button.dart';
+import '../../../../ui/weather/widgets/background_image.dart';
 
 /// [ErrorDisplayScreen]は、APIからのエラーメッセージを取得し、それを画面中央に表示する画面です。
 class ErrorDisplayScreen extends ConsumerWidget {
